requirements.txt
flask
tensorflow
numpy
pillow
requests
gtts
python-dotenv
gunicorn